[Aggregated Metrics] Reframe MetricTimePeriodJoinType (Recidiviz/recidiviz-data#36030)
## Description of the change

As I was implementing optimized metrics for PeriodSpan type metrics, I
realized that my definition of `MetricTimePeriodJoinType` ("Describes
the type of join logic we'll use to associate metric time periods with
unit of observation to unit of assignment periods.") wasn't really
describing how I was using this enum - I need an enum specifically to
tell me how to join metric periods to assignment periods with nothing
about observations involved.

This PR:
* Renames `MetricTimePeriodJoinType` to
`MetricTimePeriodToAssignmentJoinType`
* Adds two types in place of `PERIOD`: `INTERSECTION` (this is the
simple intersection join we will use for period span metrics) and
`INTERSECTION_EXTENDED` (this is the join that we use with period event
metrics)
* Renames the `event_applies_to_period_start_date` to
`intersection_start_date` - we will be able to reuse this value when
building period span metrics
* Renames `event_applies_to_period_end_date_exclusive_nonnull` to
`intersection_extended_end_date_exclusive_nonnull`
* Makes the output columns of the assignments by time period views
*variable* - we only produce date columns that will be relevant later
for the metric that is using this join type.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29291

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 36a980dbe80c354dbb825eaa0f350a07278320fa