[ID perf] add task completion metrics to generated spreadsheets (Recidiviz/recidiviz-data#36172)
## Description of the change

Adds task completions to the aggregated metrics query + puts the results
in the correct "grants during time period" cells.

I learned while doing this that if you have two enums with the same
value, python considers them aliases of each other, which means
`get_row_index` was considering them all to be the same row, which is
wrong. To handle that, I set the enum values to all be distinct using
`auto()`, and added a dict with the labels. This way, the lookup can use
the distinct enum value, but we can still write the correct heading
labels to the spreadsheet.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35935

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: de86e3ad6380ec6c45e7675d6cc6b585668a2c60