[TES] Fix default value for criteria groups with gaps (Recidiviz/recidiviz-data#36214)
## Description of the change

Fix a bug in the task criteria groups where a sub-condition with gaps
are not being filled in with the criteria default value. This issue
incorrectly sets the `meets_criteria` solely from the criteria spans
that overlap with each sub-session instead of ALL criteria
sub-conditions within the AND/OR task criteria group.

Add tests for both AND/OR criteria groups for this specific example.
Also, migrate a portion of the task criteria group tests to use the BQ
Emulator instead of testing the query template.

There are no changes to Workflows eligibility following these updates!

## Related issues

Closes Recidiviz/recidiviz-data#35726

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1a2c4ffd2b34cb5b3e63de706db6cf31ef2b1f3d