[raw data] Modify all datetime_sql_parsers to use PARSE_DATETIME (Recidiviz/recidiviz-data#36583)
## Description of the change

In BQ a TIMESTAMP is a DATETIME with timezone info. In practice, we do
not parse timezone information anywhere in the codebase when doing
PARSE_TIMESTAMP.

previously, a `datetime_sql_parser` could parse as a `DATE`, `DATETIME`,
or `TIMESTAMP`, then we cast all types to DATETIME and then back to
STRING when generating the latest view.

this results in confusing experience (in some raw data yamls we are
using `TIMESTAMP` and in others we are using `DATETIME` or `DATE` to
parse the exact same formats with no apparent reason behind when we pick
which type) and redundant casting logic. so since they are all cast to a
`DATETIME`, we should just parse them as a `DATETIME` from the
beginning.

this PR
1. Changes `datetime_sql_parsers` validation from validating that all
parsers for a column have the same type to just validating that we are
using `PARSE_DATETIME`
2. Removes the cast to DATETIME in the latest view query
3. Changes all `PARSE_DATE`/`PARSE_TIMESTAMP`s to `PARSE_DATETIME`

if we were to receive a column that had timezone information, it would
be tricky to handle because we are already making the probably incorrect
assumption that all other datetime columns are in UTC, so in some ways i
think it would be better to ignore the tz info and treat it like a UTC
datetime so its consistent with other columns. but if we were to receive
tz info and we wanted to be more accurate with our datetimes wrt
timezones, it would require some rework on how we are handling
datetimes.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35474

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant
…art from file's update_datetime

GitOrigin-RevId: c772c6f8e87b069b09d9240a1146fb4d5847c158