[ID perf] add monthly grant rate as a formula in the perf spreadsheets (Recidiviz/recidiviz-data#36623)
## Description of the change

We need to calculate the grant rate as # grants / avg caseload. Since we
already have both as rows in the sheet, we can just use a formula. The
year column is supposed to be the average of all the month values. Since
we don't have data for some newer staff members for some months, I set
the cells to empty if they were missing caseload data for that month,
and then the AVERAGE function ignores those empty cells.

Screenshot of an officer with partial data to show how it behaves:
https://drive.google.com/file/d/1RyU3jNmf4LWBlsBp7IhMghIUrMQpOVGh/view?usp=sharing

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35936

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cc3cd173c7e14b39a44b8a8dfeb1d7169720b05d